Fixed bug when stream is rotate the decode thread is end
